#7 Fixed heroku V1.0.0-Beta, Move album' songs to Song.  NO-DATABASE
